+++ x/bank/types/querier.go
@@ -35,6 +35,14 @@ func NewQuerySpendableBalancesRequest(addr sdk.AccAddress, req *query.PageReques
 	return &QuerySpendableBalancesRequest{Address: addr.String(), Pagination: req}
 }
 
+// NewQuerySpendableBalanceByDenomRequest creates a new instance of a
+// QuerySpendableBalanceByDenomRequest.
+//
+//nolint:interfacer
+func NewQuerySpendableBalanceByDenomRequest(addr sdk.AccAddress, denom string) *QuerySpendableBalanceByDenomRequest {
+	return &QuerySpendableBalanceByDenomRequest{Address: addr.String(), Denom: denom}
+}
+
 // QueryTotalSupplyParams defines the params for the following queries:
 //
 // - 'custom/bank/totalSupply'